flask-api/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
python-decouple==3.3
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
marshmallow
Werkzeug~=1.0.1
SQLAlchemy~=1.3.20
pyjwt==1.7.1
Flask-JWT-Extended==3.25.0
passlib==1.7.4